+++ bin/input-remapper-service
@@ -25,7 +25,7 @@
 import sys
 from argparse import ArgumentParser
 
-from inputremapper.logger.logger import logger
+from inputremapper.logging.logger import logger
 
 if __name__ == "__main__":
     parser = ArgumentParser()
