+++ inputremapper/configs/data.py
@@ -27,7 +27,7 @@ import sys
 
 import pkg_resources
 
-from inputremapper.logger.logger import logger
+from inputremapper.logging.logger import logger
 
 logged = False
