+++ inputremapper/gui/user_interface.py
@@ -62,7 +62,7 @@ from inputremapper.gui.utils import (
     gtk_iteration,
 )
 from inputremapper.injection.injector import InjectorStateMessage
-from inputremapper.logger.logger import logger, COMMIT_HASH, VERSION, EVDEV_VERSION
+from inputremapper.logging.logger import logger, COMMIT_HASH, VERSION, EVDEV_VERSION
 
 # https://cjenkins.wordpress.com/2012/05/08/use-gtksourceview-widget-in-glade/
 GObject.type_register(GtkSource.View)
