+++ inputremapper/injection/event_reader.py
@@ -32,7 +32,7 @@ from inputremapper.injection.mapping_handlers.mapping_handler import (
     NotifyCallback,
 )
 from inputremapper.input_event import InputEvent
-from inputremapper.logger.logger import logger
+from inputremapper.logging.logger import logger
 from inputremapper.utils import get_device_hash, DeviceHash
