+++ inputremapper/injection/mapping_handlers/abs_to_abs_handler.py
@@ -33,7 +33,7 @@ from inputremapper.injection.mapping_handlers.mapping_handler import (
     InputEventHandler,
 )
 from inputremapper.input_event import InputEvent, EventActions
-from inputremapper.logger.logger import logger
+from inputremapper.logging.logger import logger
 from inputremapper.utils import get_evdev_constant_name
