+++ inputremapper/injection/mapping_handlers/key_handler.py
@@ -29,7 +29,7 @@ from inputremapper.injection.mapping_handlers.mapping_handler import (
     HandlerEnums,
 )
 from inputremapper.input_event import InputEvent
-from inputremapper.logger.logger import logger
+from inputremapper.logging.logger import logger
 from inputremapper.utils import get_evdev_constant_name
