+++ inputremapper/injection/mapping_handlers/macro_handler.py
@@ -31,7 +31,7 @@ from inputremapper.injection.mapping_handlers.mapping_handler import (
     HandlerEnums,
 )
 from inputremapper.input_event import InputEvent
-from inputremapper.logger.logger import logger
+from inputremapper.logging.logger import logger
 
 
 class MacroHandler(MappingHandler):
