+++ inputremapper/injection/mapping_handlers/mapping_handler.py
@@ -69,7 +69,7 @@ from inputremapper.configs.input_config import InputCombination, InputConfig
 from inputremapper.configs.mapping import Mapping
 from inputremapper.exceptions import MappingParsingError
 from inputremapper.input_event import InputEvent
-from inputremapper.logger.logger import logger
+from inputremapper.logging.logger import logger
 
 
 class EventListener(Protocol):
