+++ inputremapper/injection/mapping_handlers/rel_to_abs_handler.py
@@ -47,7 +47,7 @@ from inputremapper.injection.mapping_handlers.mapping_handler import (
     InputEventHandler,
 )
 from inputremapper.input_event import InputEvent, EventActions
-from inputremapper.logger.logger import logger
+from inputremapper.logging.logger import logger
 
 
 class RelToAbsHandler(MappingHandler):
