+++ inputremapper/injection/mapping_handlers/rel_to_btn_handler.py
@@ -30,7 +30,7 @@ from inputremapper.injection.mapping_handlers.mapping_handler import (
     InputEventHandler,
 )
 from inputremapper.input_event import InputEvent, EventActions
-from inputremapper.logger.logger import logger
+from inputremapper.logging.logger import logger
 
 
 class RelToBtnHandler(MappingHandler):
