+++ inputremapper/injection/mapping_handlers/rel_to_rel_handler.py
@@ -45,7 +45,7 @@ from inputremapper.injection.mapping_handlers.mapping_handler import (
     InputEventHandler,
 )
 from inputremapper.input_event import InputEvent
-from inputremapper.logger.logger import logger
+from inputremapper.logging.logger import logger
 
 
 def is_wheel(event) -> bool:
