+++ inputremapper/ipc/pipe.py
@@ -42,7 +42,7 @@ import time
 from typing import Optional, AsyncIterator, Union
 
 from inputremapper.configs.paths import PathUtils
-from inputremapper.logger.logger import logger
+from inputremapper.logging.logger import logger
 
 
 class Pipe:
